furiosa_visa_std::prelude

Enum Opt

pub enum Opt<D> {
    Init(D),
    Uninit,
}

A data type that can be either initialized or uninitialized.

Variants

Init(D)

Initialized value.

Uninit

Uninitialized value.

Implementations

impl<D> Opt<D>

pub fn map<D2>(self, f: impl FnOnce(D) -> D2) -> Opt<D2>

Maps the initialized value using the provided function, or returns uninitialized.

pub fn zip_map<D2, R>( self, other: Opt<D2>, f: impl FnOnce(D, D2) -> R, ) -> Opt<R>

Combines two Opt values with a function. Returns Init only if both are Init.

pub fn unwrap(self) -> D

Returns the initialized value, or panics if uninitialized.