furiosa_visa_std::prelude

Trait HasBinaryOp

pub trait HasBinaryOp<D>: Clone + Copy {
    // Required method
    fn binary_op_fn(
        self,
        mode: Option<BinaryArgMode>,
    ) -> impl Fn(Opt<D>, Opt<D>) -> Opt<D>;
}

Trait for ops that provide binary operation function.

Required Methods

fn binary_op_fn( self, mode: Option<BinaryArgMode>, ) -> impl Fn(Opt<D>, Opt<D>) -> Opt<D>

Returns a function that applies this binary operation with the given mode. If mode is None, uses the default mode (Mode01).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.