furiosa_visa_std::prelude

Type Alias VectorBranchTensor

pub type VectorBranchTensor<'l, const T: Tu, D, Chip, Cluster, Slice, Time, Packet, StashD, Stash, const VE_ORDER: VeOrder, FS = Standalone, const W: PacketMode = { Way8 }> = VectorTensor<'l, T, Branch, D, Chip, Cluster, Slice, Time, Packet, StashD, Stash, VE_ORDER, FS, W>;

Tensor after branch unit.

Aliased Type

pub struct VectorBranchTensor<'l, const T: Tu, D, Chip, Cluster, Slice, Time, Packet, StashD, Stash, const VE_ORDER: VeOrder, FS = Standalone, const W: PacketMode = { Way8 }> { /* private fields */ }

Implementations

impl<'l, const T: Tu, D: VeScalar, Chip: M, Cluster: M, Slice: M, Time: M, Packet: M, const VE_ORDER: VeOrder> VectorBranchTensor<'l, T, D, Chip, Cluster, Slice, Time, Packet, D, NoTensor, VE_ORDER>

pub fn new( ctx: &'l mut TuContext<T>, inner: Tensor<D, Pair<Chip, Pair<Cluster, Pair<Slice, Pair<Time, Packet>>>>>, branch_config: BranchMode, ) -> Self

Creates a new VectorBranchTensor from inner tensor and branch configuration.